ci: build image on PR and merge
We build the image on PR and merge to master. We don't publish yet.

The image version is hardcoded for now and will match the piper version
it's built on. This is hardcoded and will need to be manually updated in
both the Dockerfile and the workflow when it changes. There's likely a
better way, but this will do for now.