ci: refactor workflows for re-usability
So that we can re-use more worflow code, we move the repeatable
processes to workflows that are triggered by `workflow_call`. A workflow
is then created for each event that triggers ci action, named after the
actual real world worflow that is happening - e.g. Create PR, Merge to
master.

We don't really need a separate workflow for building and pushing, as
the only difference is one boolean. Now both PR and Merge workflows call
build-image, passing the desired push behavior.